Clarify that docstrings are CC BY-SA 4.0 licensed
The docstrings are processed into documentation using Sphinx.  Therefore,
it can be said that the Sphinx-generated docs are derivatives of the
source.

My intention is for the docs to be licensed under CC BY-SA 4.0.  So, that
means the docstrings should also be licensed under CC BY-SA 4.0, in
addition to the BSD 3-Clause license which covers the rest of the code.