Remove version-checking stuff that doesn't work.
Unfortunately, changing the version number in a patch really isn't going to
work, because of the many ways of installing a package.  So, the pyldap
requirement is changed to one that we know is OK with the patch.

The python-ldap requirement (which can't match right now) is left alone, since
I don't have a patch for it.