Finish thread support for polling
This finishes the work needed to get polling to work in a thread.
This is extremely helpful, because it allows clients to run a main thread
which catches signals, so that the OpenLDAP client code doesn't catch those
signals and cut the connection.

We do not subclass `threading.Thread`, which was my original thought.
Instead, we provide the method `run`, a callable that can be passed as a
target to the `Thread` constructor.

Locks are already in place around the variable used to track if we should
be stopping, so that the client's main thread can call `please_stop`
safely.