Track what's been done, for cleanup mid-init
When an exception happens in initialization, we get called to
delete/unbind.  This puts mechanisms in place for us to tell if each
external thing (that is, each shelf, and the LDAP connection) is in a
state for it to be cleaned up.

Then, we only clean up the things that have been opened.

Found as part of issue #17.