Sanitize LDAP URL before connection open
The OpenLDAP client library, in its documentation for the
`ldap_initialize` function (which is what pyldap/python-ldap calls),
says that the URL passed must only contain the scheme, host, and port
components.

We need the full LDAP URL (since that’s how we get our parameters!).
So, sanitize a copy for OpenLDAP.

Fixes #17.