refresh_done callback now has item list
The `refresh_done` callback now includes a parameter: A dict-like
object which contains the entire directory.  Keys are DNs.  Each value
is a dict of attribute names & their values.

To do this, we make a bare-bones class which implements the dict
methods.  We rely on the DN-UUID map for its iterator.

The LoggingCallback has also been modified to print out the entire
directory when the `refresh_done` callback is received.