Restrict pyasn1 to ≥0.2.2 <0.3.1
pyasn1 0.3.1 and later throws an exception when pyldap/python-ldap
tries to access an optional ASN.1 member, `cookie`, that wasn’t defined
(see `syncrepl.py` around line 138).

Until this can be addressed somehow, restrict the acceptable versions
of pyasn1.  The low end of 0.2.2 is because that version has some
security fixes.

Closes #18.