Start code to switch to SQLite
Since the shelve is not able to handle as much info as we’re putting
in, start the work to switch to SQLite.

There’ll be a separate class to handle the database, with some custom
code to convert UUIDs and attribute lists into binary strings to store
in the database.

We’ll also give clients the option to use the same database as us, so
long as they’re OK with a small pile of restrictions!