Move refreshdone to DB
The ItemList class had to be completely re-created.

We now have a lazy-populating object.  When someone calls it, it
populates the list of DNs, and the number of DNs.  Attribute lists are
only loaded from the DB when necessary, and we cache all of our loads.

On the good side, I discovered the Mapping ABC, so hooray for fewer
methods to implement!