DB: Fully remove adapter
sqlite3’s register_adapter is too dangerous, because affects all
attempts to insert a type into the database, and it’s not possible to
restrict it to a certain column (as you can with register_converter.

So, we still use register_converter to de-pickle the `OBJECT` column,
but we no longer pickle automatically.