Move all optimize and vacuum runs to Syncrepl
Although our db code knows that threading can be an issue, it’s the
Syncrepl code where threading-related reconnects are performed.  Since
we were doing an optimize run at db destruction, change that.

We now expose optimize and vacuum as methods to call.  Syncrepl calls
vacuum after refresh is done, and optimize on connect, on unbind, and
right before vacuum.

We can probably also call optimize occasionally through the run, but
we’d need to track that somehow.