Specifically check for Python major jumps
Our original method for checking Python versions breaks when moving
from Python 2 to Python 3.  It’s mainly because we’re storing our DB
setting as pickled objects, which has a different issue depending on
the direction we’re trying to move:

• From 2 to 3, 3 can read the data, but complains about strings vs.
bytes.

• From 3 to 2, 2 can’t read the pickled data from 3 (there’s a newer
format).

So, we store the major version in its own one-column table.  If there’s
a mismatch, then we fail before reading anything else.  If the match is
OK, then we continue as normal.  This assumed that there won’t be any
major breaks _within_ a Python major version.