Use SQLite WAL mode with NORMAL sync
Informal benchmarking shows that, at least in the refresh phase, using
the WAL journal can dramatically speed up the database.  But, to
maintain durability, we use the NORMAL synchronous mode.  We also do a
WAL checkpoint at the end of the refresh phase.

TODO: Do a checkpoint (and an optimize) from time to time, when in the
persist phase.